feat: add utility function to add course by URL
