+++ src/views/components/injected/AddAllButton.tsx
@@ -1,5 +1,6 @@
 import { background } from '@shared/messages';
 import type { Course } from '@shared/types/Course';
+import { addCourseByUrl } from '@shared/util/courseUtils';
 import { Button } from '@views/components/common/Button';
 import ExtensionRoot from '@views/components/common/ExtensionRoot/ExtensionRoot';
 import useSchedules from '@views/hooks/useSchedules';
@@ -21,21 +22,21 @@ export default function InjectedButton(): JSX.Element | null {
     const [container, setContainer] = useState<HTMLDivElement | null>(null);
     const [activeSchedule, _] = useSchedules();
 
-    const removeDuplicates = (courseIds: string[], courses: string[]) => {
-        const uniqueCourseIds = [] as string[];
-        const uniqueCourses = [] as string[];
-        const seen = new Set();
+    // const removeDuplicates = (courseIds: string[], courses: string[]) => {
+    //     const uniqueCourseIds = [] as string[];
+    //     const uniqueCourses = [] as string[];
+    //     const seen = new Set();
 
-        courseIds.forEach((id, index) => {
-            if (!seen.has(id)) {
-                seen.add(id);
-                uniqueCourseIds.push(id);
-                uniqueCourses.push(courses[index] as string);
-            }
-        });
+    //     courseIds.forEach((id, index) => {
+    //         if (!seen.has(id)) {
+    //             seen.add(id);
+    //             uniqueCourseIds.push(id);
+    //             uniqueCourses.push(courses[index] as string);
+    //         }
+    //     });
 
-        return { uniqueCourseIds, uniqueCourses };
-    };
+    //     return { uniqueCourseIds, uniqueCourses };
+    // };
 
     const extractCoursesFromCalendar = () => {
         const calendarElement = document.querySelector('#kgoui_Rcontent_I3_Rprimary_I1_Rcontent_I1_Rcontent_I0_Ritems');
@@ -46,38 +47,47 @@ export default function InjectedButton(): JSX.Element | null {
         }
 
         const anchorTags = Array.from(calendarElement.querySelectorAll('a')).filter(
-            anchor => !anchor.href.includes('google.com') // Remove the Google Map links
+            anchor => !anchor.href.includes('google.com')
         );
 
-        const courses = Array.from(anchorTags).map(x => x.innerText.trim());
-        const courseIds = Array.from(anchorTags).map(({ href }) => {
-            const segments = href.split('/').filter(Boolean); // Remove empty segments caused by trailing slashes
-            return segments.pop();
-        }) as string[];
+        // Make sure to remove duplicate anchorTags using set
+        const uniqueAnchorTags = new Set(anchorTags);
 
-        const { uniqueCourseIds, uniqueCourses } = removeDuplicates(courseIds, courses);
-
-        handleAddingCourses(uniqueCourses, uniqueCourseIds);
-    };
-
-    const handleAddingCourses = async (courses: string[], uniqueCourseIds: string[]) => {
-        courses.forEach(async (course, index) => {
-            // TODO: need to still get the other course data
-            const courseData = {
-                id: uniqueCourseIds[index],
-                name: course,
-            } as unknown as Course;
-
-            try {
-                await addCourse({
-                    course: courseData,
-                    scheduleId: activeSchedule.id,
-                });
-            } catch (error) {
-                console.error(error);
-            }
+        uniqueAnchorTags.forEach(anchor => {
+            addCourseByUrl(anchor.href, activeSchedule);
         });
+
+        // const courses = Array.from(anchorTags).map(x => x.innerText.trim());
+        // const courseIds = Array.from(anchorTags).map(({ href }) => {
+        //     const segments = href.split('/').filter(Boolean); // Remove empty segments caused by trailing slashes
+        //     return segments.pop();
+        // }) as string[];
+
+        // const { uniqueCourseIds, uniqueCourses } = removeDuplicates(courseIds, courses);
+
+        // // print all the href of anchor tags
+        // console.log(Array.from(anchorTags).map(x => x.href));
+        // handleAddingCourses(uniqueCourses, uniqueCourseIds);
     };
+
+    // const handleAddingCourses = async (courses: string[], uniqueCourseIds: string[]) => {
+    //     courses.forEach(async (course, index) => {
+    //         // TODO: need to still get the other course data
+    //         const courseData = {
+    //             id: uniqueCourseIds[index],
+    //             name: course,
+    //         } as unknown as Course;
+
+    //         try {
+    //             await addCourse({
+    //                 course: courseData,
+    //                 scheduleId: activeSchedule.id,
+    //             });
+    //         } catch (error) {
+    //             console.error(error);
+    //         }
+    //     });
+    // };
 
     useEffect(() => {
         const targetElement = document.getElementById('kgoui_Rcontent_I3_Rsecondary');
